chore(upsell-serializer): move transformBlocks to utisl file
